+++ packages/grid/src/features/standard-table/stories/StandardTable.stories.tsx
@@ -1,4 +1,3 @@
-import { Store, withState } from "@dump247/storybook-state";
 import {
   createColumnConfig,
   createEditableTextCellWithStatus,
@@ -9,7 +8,7 @@ import {
 import { addDays, format } from "date-fns";
 import * as React from "react";
 import { useCallback, useState } from "react";
-import { Box, Column, Row, Spacing, StandardText } from "@stenajs-webui/core";
+import { Box, Column, Spacing, StandardText } from "@stenajs-webui/core";
 
 interface ListItem {
   id: string;
@@ -70,7 +69,7 @@ const createItemsMocks = (): Array<ListItem> => [
     departure: addDays(new Date(), 14),
   },
   {
-    id: "129",
+    id: "A really long id, or could be a very long title",
     active: false,
     name: "Schenker",
     ship: "MS Britannica",
@@ -91,40 +90,38 @@ const standardTableConfigForStories: StandardTableConfig<ListItem> = {
   showRowCheckbox: true,
   enableGridCell: true,
   columns: {
-    id: createColumnConfig((item) => item.id),
+    id: createColumnConfig((item) => item.id, {
+      width: "100px",
+    }),
     active: createColumnConfig((item) => item.active, {
       itemLabelFormatter: (value) => (value ? "Y" : ""),
       infoIconTooltipText: "Active means out on the sea.",
+      width: "150px",
     }),
-    name: createColumnConfig((item) => item.name),
-    ship: createColumnConfig((item) => item.ship),
+    name: createColumnConfig((item) => item.name, {
+      width: "200px",
+    }),
+    ship: createColumnConfig((item) => item.ship, {
+      width: "150px",
+    }),
     numPassengers: createColumnConfig((item) => item.numPassengers, {
       renderCell: createStandardEditableTextCell(),
       isEditable: true,
       onChange: () => {},
       justifyContentHeader: "flex-end",
       justifyContentCell: "flex-end",
+      width: "200px",
     }),
     departure: createColumnConfig((item) => item.departure, {
       itemLabelFormatter: (value) => format(value, "yyyy-MM-dd"),
       borderLeft: true,
+      width: "100px",
     }),
   },
   columnOrder: ["id", "active", "name", "ship", "numPassengers", "departure"],
 };
 
 const mockedItems = createItemsMocks();
-
-const createOnChangeNumPassengers = (
-  store: Store<{ items: Array<ListItem> }>
-) => (item: ListItem, numPassengers: string | undefined) => {
-  const items = setListItemFields(store.state.items, item.id, {
-    numPassengers: numPassengers ? parseInt(numPassengers) : undefined,
-  });
-  return store.set({
-    items,
-  });
-};
 
 export default {
   title: "grid/StandardTable",
@@ -167,124 +164,196 @@ export const Standard = () => {
   return <StandardTable items={items} config={config} />;
 };
 
-export const WithSortingDisabled = withState({ items: mockedItems })(
-  ({ store }) => {
-    const config = createConfig(
-      createOnChangeNumPassengers(store),
-      createStandardEditableTextCell(),
-      { disableSorting: true }
-    );
-    return <StandardTable items={store.state.items} config={config} />;
-  }
-);
+export const WithSortingDisabled = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
 
-export const WithSortingDisabledAndSortByName = withState({
-  items: mockedItems,
-})(({ store }) => {
-  const config = createConfig(
-    createOnChangeNumPassengers(store),
-    createStandardEditableTextCell(),
-    { disableSorting: true, initialSortOrder: "name" }
-  );
-  return <StandardTable items={store.state.items} config={config} />;
-});
+  const config: StandardTableConfig<ListItem> = {
+    disableSorting: true,
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
 
-export const WithSortingEnabledAndSortByNameDesc = withState({
-  items: mockedItems,
-})(({ store }) => {
-  const config = createConfig(
-    createOnChangeNumPassengers(store),
-    createStandardEditableTextCell(),
-    {
-      initialSortOrder: "name",
-      initialSortOrderDesc: true,
-    }
-  );
-  return <StandardTable items={store.state.items} config={config} />;
-});
+  return <StandardTable items={items} config={config} />;
+};
 
-export const WithFieldError = withState({ items: mockedItems })(({ store }) => {
-  const config = createConfig(
-    createOnChangeNumPassengers(store),
-    createEditableTextCellWithStatus<number | undefined, ListItem>(
-      undefined,
-      (item) => ({
-        hasError: true,
-        errorMessage: "Something failed.",
-        id: item.id,
-      })
-    )
-  );
-  return <StandardTable items={mockedItems} config={config} />;
-});
+export const WithSortingDisabledAndSortByName = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
 
-export const WithFieldLoading = withState({ items: mockedItems })(
-  ({ store }) => {
-    const config = createConfig(
-      createOnChangeNumPassengers(store),
-      createEditableTextCellWithStatus<number | undefined, ListItem>(
-        undefined,
-        (item) => ({
+  const config: StandardTableConfig<ListItem> = {
+    disableSorting: true,
+    initialSortOrder: "name",
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const WithSortingEnabledAndSortByNameDesc = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    initialSortOrder: "name",
+    initialSortOrderDesc: true,
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const WithFieldError = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        renderCell: createEditableTextCellWithStatus<
+          number | undefined,
+          ListItem
+        >(undefined, (item) => ({
+          hasError: true,
+          errorMessage: "Something failed.",
+          id: item.id,
+        })),
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const WithFieldLoading = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        renderCell: createEditableTextCellWithStatus<
+          number | undefined,
+          ListItem
+        >(undefined, (item) => ({
           id: item.id,
           loading: true,
-        })
-      )
-    );
-    return <StandardTable items={mockedItems} config={config} />;
-  }
+        })),
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const WithModifiedFields = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        renderCell: createEditableTextCellWithStatus<
+          number | undefined,
+          ListItem
+        >(
+          "Passengers cannot be empty",
+          () => undefined,
+          (item) => ({
+            id: item.id,
+            modified: true,
+            newValue: "789",
+          })
+        ),
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const WithWarningWhenModifiedFieldIsEmpty = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        renderCell: createEditableTextCellWithStatus<
+          number | undefined,
+          ListItem
+        >(
+          "Passengers cannot be empty",
+          () => undefined,
+          (item) => ({
+            id: item.id,
+            modified: true,
+            newValue: "",
+          })
+        ),
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
+};
+
+export const MissingItems = () => (
+  <StandardTable items={[]} config={standardTableConfigForStories} />
 );
-
-export const WithModifiedFields = withState({ items: mockedItems })(
-  ({ store }) => {
-    const config = createConfig(
-      createOnChangeNumPassengers(store),
-      createEditableTextCellWithStatus<number | undefined, ListItem>(
-        "Passengers cannot be empty.",
-        () => undefined,
-        (item) => ({
-          id: item.id,
-          modified: true,
-          newValue: "789",
-        })
-      )
-    );
-    return <StandardTable items={mockedItems} config={config} />;
-  }
-);
-
-export const WithWarningWhenModifiedFieldIsEmpty = withState({
-  items: mockedItems,
-})(({ store }) => {
-  const config = createConfig(
-    createOnChangeNumPassengers(store),
-    createEditableTextCellWithStatus<number | undefined, ListItem>(
-      "Passengers cannot be empty.",
-      () => undefined,
-      (item) => ({
-        id: item.id,
-        modified: true,
-        newValue: "",
-      })
-    )
-  );
-  return <StandardTable items={mockedItems} config={config} />;
-});
-
-export const MissingItems = () => <StandardTable items={[]} config={config} />;
 
 export const NavigationBetweenTables = () => {
-  const config = createConfig(undefined, undefined, {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
     gridCellOptions: {
       edgeMode: "unlimited",
     },
-  });
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
 
   return (
     <Column>
-      <StandardTable items={mockedItems} config={config} tableId={"table123"} />
+      <StandardTable items={items} config={config} tableId={"table123"} />
       <Spacing />
       <StandardTable
-        items={mockedItems}
+        items={items}
         config={config}
         tableId={"table123"}
         rowIndexOffset={mockedItems.length}
@@ -295,32 +364,48 @@ export const NavigationBetweenTables = () => {
 };
 
 export const Loading = () => (
-  <StandardTable items={mockedItems} config={config} loading />
+  <StandardTable
+    items={mockedItems}
+    config={standardTableConfigForStories}
+    loading
+  />
 );
 
 export const _Error = () => (
   <StandardTable
     items={mockedItems}
-    config={config}
+    config={standardTableConfigForStories}
     error={new Error("Could not fetch users")}
   />
 );
 
 export const ExpandableRows = () => {
-  const config = createConfig(undefined, undefined, {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
     enableExpandCollapse: true,
-    showHeaderExpandCollapse: true,
     renderRowExpansion: (item) => (
       <Box spacing indent>
         <StandardText>Name: {item.name}</StandardText>
       </Box>
     ),
-  });
-  return <StandardTable items={mockedItems} config={config} />;
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
 };
 
 export const SomeExpandableRows = () => {
-  const config = createConfig(undefined, undefined, {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
     enableExpandCollapse: true,
     expandCollapseDisableResolver: (item) =>
       item.numPassengers != null && item.numPassengers > 500,
@@ -329,38 +414,109 @@ export const SomeExpandableRows = () => {
         <StandardText>Name: {item.name}</StandardText>
       </Box>
     ),
-  });
-  return <StandardTable items={mockedItems} config={config} />;
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return <StandardTable items={items} config={config} />;
 };
 
 export const WithStickyTableHeader = () => {
-  const config = createConfig(undefined, undefined, {
-    stickyHeader: true,
-  });
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    headerRowSticky: true,
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
   return (
     <Box style={{ maxHeight: "220px", overflowY: "scroll" }}>
-      <StandardTable items={mockedItems} config={config} />
+      <StandardTable items={items} config={config} />
     </Box>
   );
 };
 
-export const MultipleTablesWithStickyTableHeader = () => {
-  const config = createConfig(undefined, undefined, {
-    stickyHeader: true,
-  });
+export const WithStickyTableHeaderConfiguration = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    headerRowSticky: true,
+    headerRowOffsetTop: 16,
+    headerRowZ: 499,
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
   return (
-    <Row
-      style={{
-        maxHeight: "220px",
-        overflowY: "scroll",
-      }}
-    >
-      <Box flex={1}>
-        <StandardTable items={mockedItems} config={config} />
+    <Box style={{ maxHeight: "220px", overflowY: "scroll" }}>
+      <Box
+        style={{
+          height: 16,
+          position: "sticky",
+          top: 0,
+          zIndex: 500,
+          backgroundColor: "white",
+        }}
+      >
+        <StandardText>
+          Other upper sticky content that should stick above header
+        </StandardText>
       </Box>
-      <Box flex={1}>
-        <StandardTable items={mockedItems} config={config} />
+      <StandardTable items={items} config={config} />
+    </Box>
+  );
+};
+
+export const WithStickyFirstColumn = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem> = {
+    ...standardTableConfigForStories,
+    showHeaderCheckbox: false,
+    showRowCheckbox: false,
+    headerRowSticky: true,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      id: {
+        ...standardTableConfigForStories.columns.id,
+        width: "245px",
+        background: "white",
+        sticky: true,
+        shadowBorder: true,
+        zIndex: 444,
+      },
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+    columnOrder: ["id", "active", "name", "ship", "numPassengers", "departure"],
+  };
+
+  return (
+    <Box style={{ maxHeight: "200px", maxWidth: "80%", overflow: "scroll" }}>
+      <Box style={{ width: "100vw" }}>
+        <StandardTable items={items} config={config} />
       </Box>
-    </Row>
+    </Box>
   );
 };