+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
+
 ## 2.2.1
+
+### New components
+
+#### SidebarMenu
+
+A menu with a title and a wrapping links and collapsible menu components
+
+- SidebarMenuContainer, a wrapper for alla `SidebarMenus` to group them vertically
+- SidebarMenuHeading, a title to the `SidebarMenu`
+- SidebarMenuLink, a link that can be selected and have an icon to the left
+- SidebarMenuCollapsible, a component that expands and contracts sub menu items `SidebarMenuCollapsibleLink`
+- SidebarMenuCollapsibleLink, a link that can be selected and is contained in a `SidebarMenuCollapsible`
 
 ### `StandardTable`