+++ CHANGELOG.md
@@ -17,6 +17,7 @@ A menu with a title and a wrapping links and collapsible menu components
 ### `StandardTable`
 
 Added option for having a sticky header row as option in `StandardTable`
+Added option for sticky first column in `StandardTable`
 
 ## 2.2.0